ui(stepper): update to take into account the latest design changes
