+++ ui-lib/src/components/ProgressSteps/Item.stories.tsx
@@ -1,19 +1,17 @@
 import React from "react";
-import { Step, StepProps } from "@components/ProgressBar";
+import { Step, StepProps } from "@components/ProgressSteps";
 
 export default {
-  title: "Navigation/ProgressBars/SidePanel/Item",
+  title: "Navigation/Stepper/SidePanel/Item",
   component: Step,
 };
 
 export const Item = (args: StepProps): JSX.Element => (
-  <div style={{ width: "75px" }}>
+  <div style={{ width: "150px" }}>
     <Step {...args} />
   </div>
 );
 Item.args = {
   state: "current",
   label: "label",
-  hideLeftSeparator: true,
-  nextState: undefined,
 };
